[Gold V] Title: 체크박스 누르기, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Gold/33931. 체크박스 누르기/체크박스 누르기.cc
@@ -4,12 +4,12 @@ using namespace std;
 
 int main()
 {
-    int N, M;
+    long long N, M;
     cin >> N >> M;
     
-    int a = M/N;
-    int b = M%N;
+    long long a = M/N;
+    long long b = M%N;
     
-    int answer = (a%2 == 0 ? b : N-b);
+    long long answer = (a%2 == 0 ? b : N-b);
     cout << answer;
 }
